scylla: add simple arg parser
For easy parsing of the api command and its args.

Although https://docs.python.org/3/library/argparse.html
provides subparsers, those has to be added in advance
to parsing the command line args, but in our case
we first want to allow setting the address and port
first so we can get the api commands, and only
then we can setup the subparsers.

The simple argparser stop at '--' or the first
unrecognized argument, placing the trailing args in
parser.extra_args.

This form the command, followed by its arguments.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>